Fixing issue where Into<DataFrame> can fail
If a multi-pin test (generating an MPR) has an execution count
less than the total part count, then the conversion into the
fixed-length array-type column in the DataFrame failed because
the underlying Vec<Vec<f32>> structure was not perfectly rectangular.

Now the multi-pin data is padded to be rectangular after parsing.